// FrontEnd/src/app/models/customer.model.ts
export class Customer {
  firstName: string;
  lastName: string;
  email: string;
  mobile: number;
  country: string;
  city: string;
  address: string;
  dateOfBirth: Date;
  internetMainUse: string;
  sex: string;
  joined: Date;
  username: String;
  password: String;
  userInformation: String;
}